refactor: move cliConstant to constant folder
