+++ src/constant/cliConstants.ts
@@ -1,4 +1,4 @@
-export const TO_DEFAULT_VALUE = 'HEAD'
+export const TO_DEFAULT_VALUE = ''
 export const OUTPUT_DEFAULT_VALUE = './output'
 export const SOURCE_DEFAULT_VALUE = './'
 export const REPO_DEFAULT_VALUE = './'
